fix: fixed bug with relative symlinking
Fixed bug introduced in the modularity goal of the project; the new
function `logged_ln` did not log the correct path for the newly created
symlink, instead referring to the directory in which the symlink will be
created. Algorithm was adjusted to log the file-to-symlink combined with
the target directory.

- Fixed bug in the `logged_ln` function
- Modified help message slightly to be more clear
- Renamed function `create_dir_if_not_exist` to
  `create_dir_if_not_exists`
- Added a brute-force check in the install function to determine whether
  or not the directory has any folders in it at all (no folders mean no
  applications)
- Compressed help check into one line as well as the number of arguments
  checker
- Changed several error messages to be clearer to the user