Added more utility functions pre-more functionality
Added initial functions in anticipation for supporting better redundancy
of installation. The user should not receive warnings when they re-run
install on everything that they've already installed. I've added the
function `file_to_array` and `string_in_array` to facilitate the reading
of cache files which will then be loaded with updated content and
checked against when installing new files.

- Added the `file_to_array` function
- Added the `string_in_array` function
- Updated the docstring of `logged_ln` to better reflect the
  functionality of that function