fix: renamed CREATED_LINKS_* to CREATED_FILES_*
Renamed anything that starts with CREATED_LINKS_* to CREATED_FILES_*.
Also, renamed the file named ".created_links" to ".created_files".

This terminology, while now a bit more terse, has been updated to
reflect the changing function of the cache file for links to a general
cache file for both links and normal files in preparation for the
ability to change installation methods.