fix: fixed bug introduced by globbing change
Fixed bug introduced with globbing changes. While the globbing utility
function that I created is great, it comes with the drawback of changing
the semantics of the path that it returns, so it's hard to match up with
the .ddmignore and requires adjusting to work correctly. I may consider
adjusting the helper function itself in the future to make this
implementation cleaner.

Also, updated tests for file globbing function.