fix: fixed regression introduced by in_array optimization
Fixed regression caused by incorrect construction of a regex.
Previously, the regex wrongly required that a space be present before
the item to be matched. This means that the regex could never match an
element that occurs at the beginning of an array because such an element
would never contain a space at the beginning.