fix: added additional docs to is_ignored
Currently is_ignored relies on previously undocumented functionality
present in the starts_with function. As it currently stands, and change
to the underlying functionality of the starts_with function is at risk
of disrupting how this function works, breaking encapsulation. This
change doesn't fix the underlying issue, but rather indicates that this
hidden reliance exists.